make quirk types seperate classes & move to quirks.py
