+++ quirks.py
@@ -1,108 +1,198 @@
 import os
+import re
 import logging
 
 import ostools
+from mispeller import mispeller
+from parsetools import parseRegexpFunctions
 
 _datadir = ostools.getDataDir()
 PchumLog = logging.getLogger("pchumLogger")
 
 
-class ScriptQuirks:
-    def __init__(self):
-        self._datadir = ostools.getDataDir()
-        self.home = os.getcwd()
-        self.quirks = {}
-        self.last = {}
-        self.scripts = []
-        # self.load()
+# traditional quirks
 
-    def loadModule(self, name, filename):
-        raise Exception
 
-    def modHas(self, module, attr):
-        return False
+def PesterQuirkFactory(quirk: dict):
+    """Returns a valid PesterQuirk object from the given quirk dictionary"""
+    # This is a "factory" because a lot of old code depends on calling the old class which was all quirks rolled into 1
+    match quirk["type"]:
+        case "prefix":
+            return PrefixPesterQuirk(quirk)
+        case "suffix":
+            return SuffixPesterQuirk(quirk)
+        case "replace":
+            return ReplacePesterQuirk(quirk)
+        case "regexp":
+            return RegexpPesterQuirk(quirk)
+        case "random":
+            return RandomPesterQuirk(quirk)
+        case "spelling":
+            return MispellerPesterQuirk(quirk)
 
-    def loadAll(self):
-        self.last = self.quirks.copy()
-        self.quirks.clear()
-        for script in self.scripts:
-            PchumLog.info(script.getExtension())
-            script.load()
-            # print script.quirks
-            for q in script.quirks:
-                self.quirks.update(script.quirks)
-        for k in self.last:
-            if k in self.quirks:
-                if self.last[k] == self.quirks[k]:
-                    del self.quirks[k]
-        # print self.quirks
-        if hasattr(self, "quirks"):
-            # See https://stackoverflow.com/questions/12843099/python-logging-typeerror-not-all-arguments-converted-during-string-formatting
-            reg_quirks = ("Registered quirks:", "(), ".join(self.quirks) + "()")
-            PchumLog.info(reg_quirks)
+
+class PesterQuirk:
+    def __init__(self, quirk: dict):
+        self.quirk = quirk
+        self.type = self.quirk["type"]
+        self.on = self.quirk.get("on", True)
+        self.group = self.quirk.get("group", "Miscellaneous")
+
+        self.checkstate = self.quirk.get(
+            "checkstate", 0
+        )  ## Seems to be somethign related to the QT checkbox? QtCore.QT.CheckState
+
+    def apply(self, string: str, first: bool = False, last: bool = False):
+        """string: string to operate quirk on. first: is the given substring at the very start (idx == 0) of the superstring? last: is the given substring at the very last (idx == -1) of the superstring?"""
+        if self.on:
+            return self._apply(string, first, last)
         else:
-            PchumLog.warning("Couldn't find any script quirks")
+            return string
 
-    def add(self, script):
-        self.scripts.append(script)
+    def _apply(self, string: str, first: bool, last: bool):
+        # Overwrite
+        raise NotImplemented()
+        return string
 
-    def load(self):
-        self.last = self.quirks.copy()
-        self.quirks.clear()
-        try:
-            extension = self.getExtension()
-        except AttributeError:
-            PchumLog.exception(
-                "No self.getExtension(), does ScriptQuirks need to be subclassed?"
-            )
-            return
-        filenames = []
-        if not os.path.exists(os.path.join(self.home, "quirks")):
-            os.makedirs(os.path.join(self.home, "quirks"), exist_ok=True)
-        for fn in os.listdir(os.path.join(self.home, "quirks")):
-            if fn.endswith(extension) and not fn.startswith("_"):
-                filenames.append(os.path.join(self.home, "quirks", fn))
-        if hasattr(self, "_datadir"):
-            if not os.path.exists(os.path.join(self._datadir, "quirks")):
-                os.makedirs(os.path.join(self._datadir, "quirks"), exist_ok=True)
-            for fn in os.listdir(os.path.join(self._datadir, "quirks")):
-                if fn.endswith(extension) and not fn.startswith("_"):
-                    filenames.append(os.path.join(self._datadir, "quirks", fn))
+    def __str__(self):
+        # Overwrite
+        return "UNKNOWN QUIRK"
 
-        modules = []
-        for filename in filenames:
-            try:
-                extension_length = len(self.getExtension())
-            except AttributeError:
-                PchumLog.exception(
-                    "No self.getExtension(), does ScriptQuirks need to be subclassed?"
-                )
-                return
-            name = os.path.basename(filename)[:-extension_length]
-            try:
-                module = self.loadModule(name, filename)
-                if module is None:
-                    continue
-            except Exception as e:
-                PchumLog.warning(
-                    "Error loading %s: %s (in quirks.py)", os.path.basename(name), e
-                )
+
+class PrefixPesterQuirk(PesterQuirk):
+    def __init__(self, quirk: dict):
+        assert quirk["type"] == "prefix"
+        super().__init__(quirk)
+
+    def _apply(self, string: str, first: bool, last: bool):
+        return self.quirk["value"] + string
+
+    def __str__(self):
+        return "BEGIN WITH: %s" % (self.quirk["value"])
+
+
+class SuffixPesterQuirk(PesterQuirk):
+    def __init__(self, quirk: dict):
+        assert quirk["type"] == "suffix"
+        super().__init__(quirk)
+
+    def _apply(self, string: str, first: bool, last: bool):
+        return string + self.quirk["value"]
+
+    def __str__(self):
+        return "END WITH: %s" % (self.quirk["value"])
+
+
+class ReplacePesterQuirk(PesterQuirk):
+    def __init__(self, quirk: dict):
+        assert quirk["type"] == "replace"
+        super().__init__(quirk)
+
+    def _apply(self, string: str, first: bool, last: bool):
+        return string.replace(self.quirk["from"], self.quirk["to"])
+
+    def __str__(self):
+        return "REPLACE {} WITH {}".format(self.quirk["from"], self.quirk["to"])
+
+
+class RegexpPesterQuirk(PesterQuirk):
+    def __init__(self, quirk: dict):
+        assert quirk["type"] == "regexp"
+        super().__init__(quirk)
+
+    def _apply(self, string: str, first: bool, last: bool):
+        # regex string
+        from_ = self.quirk["from"]
+
+        # Exit prematurely if the regexp is only supposed to act on the first substring of the superstring and this isnt that (^ is start of string)
+        if not first and len(from_) > 0 and from_[0] == "^":
+            return string
+        # Exit prematurely if the regexp is only supposed to act on the last substring of the superstring and this isnt that ($ is end of string)
+        if not last and len(from_) > 0 and from_[-1] == "$":
+            return string
+
+        # the replace string
+        to = self.quirk["to"]
+        # I think this handles the regex functions like rainbow()
+        parse_tree = parseRegexpFunctions(to)
+        return re.sub(from_, parse_tree.expand, string)
+
+    def __str__(self):
+        return "REGEXP: {} REPLACED WITH {}".format(
+            self.quirk["from"],
+            self.quirk["to"],
+        )
+
+
+class RandomPesterQuirk(PesterQuirk):
+    def __init__(self, quirk: dict):
+        assert quirk["type"] == "random"
+        super().__init__(quirk)
+
+    def _apply(self, string: str, first: bool, last: bool):
+        # Fallback if the quirk is not set up right (no random strings to replace with)
+        if len(self.quirk.get("randomlist", [])) == 0:
+            return string
+
+        # regex string
+        from_ = self.quirk["from"]
+
+        # See regexPesterQuirk
+        if not first and len(from_) > 0 and from_[0] == "^":
+            return string
+        if not last and len(from_) > 0 and from_[-1] == "$":
+            return string
+
+        # Pick random item
+        # I believe this gets called for each match in the re.sub
+        def randomrep(mo):
+            choice = random.choice(self.quirk["randomlist"])
+            parse_tree = parseRegexpFunctions(choice)
+            return parse_tree.expand(mo)
+
+        return re.sub(from_, randomrep, string)
+
+    def __str__(self):
+        return "REGEXP: {} RANDOMLY REPLACED WITH {}".format(
+            self.quirk["from"],
+            self.quirk["randomlist"],
+        )
+
+
+class MispellerPesterQuirk(PesterQuirk):
+    def __init__(self, quirk: dict):
+        assert quirk["type"] == "spelling"
+        super().__init__(quirk)
+
+    def _apply(self, string: str, first: bool, last: bool):
+        percentage = self.quirk["percentage"] / 100.0
+        out = []
+        # regex to avoid color tags
+        ctag = re.compile("(</?c=?.*?>)", re.I)
+
+        # Split by space to get all words in given string
+        for word in string.split(" "):
+            # get random 0.0 - 1.0 number
+            dice = random.random()
+
+            if not ctag.search(w) and dice < percentage:
+                # word is not wrapped in color tags :)
+                out.append(mispeller(word))
+            elif dice < percentage:
+                # word IS wrapped in color tags!!
+                tmp = []
+                split = ctag.split(word)
+                # Only garble substrings if they are not a <c> tag
+                for sequence in split:
+                    if sequence and not ctag.search(sequence):
+                        tmp.append(mispeller(sequence))
+                    else:
+                        tmp.append(sequence)
+                out.append("".join(tmp))
             else:
-                if self.modHas(module, "setup"):
-                    module.setup()
-                if self.modHas(module, "commands"):
-                    self.register(module)
-                modules.append(name)
-        for k in self.last:
-            if k in self.quirks:
-                if self.last[k] == self.quirks[k]:
-                    del self.quirks[k]
+                out.append(word)
+        # Turn back into normal sentence
+        return " ".join(out)
 
-    def funcre(self):
-        if not hasattr(self, "quirks"):
-            return r"\\[0-9]+"
-        f = r"("
-        for q in self.quirks:
-            f = f + q + r"\(|"
-        f = f + r"\)|\\[0-9]+)"
-        return f
+    def __str__(self):
+        return "MISPELLER: %d%%" % (self.quirk["percentage"])
